feat(frontend): FE-6 Server Components + metadata
Make app/layout.tsx a Server Component with a Metadata export, moving all
client runtime (Redux/persist, MetaMask SDK, navbar, banner, toasts) into a
new Providers client boundary. Convert read-only about/terms/privacy to
static Server Components with metadata, and add a server page.tsx wrapper for
/products/[id] with generateMetadata (OG image/title/description) while
keeping its interactive detail UI as a client island.
