+++ app/privacy/page.tsx
@@ -1,7 +1,12 @@
-"use client";
-
+import type { Metadata } from "next";
 import InfoPage, { InfoSection } from "@/components/ui/info-page";
 import Link from "next/link";
+
+// FE-6: read-only page → Server Component with SEO metadata.
+export const metadata: Metadata = {
+  title: "Privacy Policy",
+  description: "The Bazaar privacy policy.",
+};
 
 const LAST_UPDATED = "May 31, 2026";
